Corrected bug in getOffset(). Without the proper cast to long from int the calculation overflows and can result in a negative offset returned

+++ src/main/java/org/mellowtech/core/io/AbstractBlockFile.java
@@ -233,7 +233,7 @@ abstract class AbstractBlockFile implements RecordFile {
   }
 
   protected long getOffset(int record) {
-    return blocksOffset() + (record * (blockSize));
+    return blocksOffset() + ((long)record * (blockSize));
   }
 
   protected long headerOffset() {
